Make datbase use given ID
When a new entity is inserted into the database, if its field ID is set
(meaning its not 0), that same ID will be used to store the entity.

If the ID is not provided (meaning it is 0), a new id will be generated.

If the given ID is already in the database, an exception in thrown.